Moved away dep from libmodule_mem in libmodule_utils_internal.
