+++ Lib/core/globals.h
@@ -11,6 +11,13 @@
 #include "mem.h"
 #include "log.h"
 
+#define M_MEM_LOCK(mem, func) \
+    m_mem_ref(mem); \
+    func; \
+    m_mem_unref(mem);
+
+void mem_dtor(void *src);
+
 /* Global variables are defined in main.c */
 extern m_map_t *ctx;
 extern pthread_mutex_t mx;          // Used to access/modify global ctx map
