+++ Lib/utils/mem.h
@@ -2,13 +2,7 @@
 
 #include <stddef.h>
 
-#define M_MEM_LOCK(mem, func) \
-    m_mem_ref(mem); \
-    func; \
-    m_mem_unref(mem);
-
 char *mem_strdup(const char *s);
-void mem_dtor(void *src);
 
 /* Struct that holds user defined memory functions */
 typedef struct {
