+++ TODO.md
@@ -71,7 +71,7 @@ Maybe try to m_mod_load() the newly created file?
 - [x] Split some more private headers from priv.h
 - [x] split libmodule_structs, libmodule_thpool, libmodule_mem ...
 - [ ] Fix pkg-config
-- [ ] test
+- [ ] test with installed lib
 
 - [x] itr.h allow const types too
